requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-modal-forms==2.2.0
django-extensions==3.1.5
django-on-heroku==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
